Update README.md
The weather application was created online May 25, 2025 as the Master AI for Web App Development course. In the creation or process of coding, ChatGPT was used by generating several questions to get the code results. Much faster than the traditional way of coding an application. I found that debugging and fixing was still needed because of the mistakes CHatGPT made. In addition, since this was a class i did not order the OpenWeatherMap API key to see if the search of states and the weather there were found.